refactor: blogs 컴포넌트 통합
