+++ app/blogs/[slug]/page.tsx
@@ -4,8 +4,8 @@ import { Box, Flex, Title } from '@mantine/core';
 import fs from 'fs';
 import path from 'path';
 import Comments from '@/components/blogs/Comments';
-import { MdxComponents } from '@/components/mdx/mdxComponents/MdxComponents';
-import MDXContent from '@/components/mdx/MDXContent';
+import { MdxComponents } from '@/components/blogs/mdxComponents/MdxComponents';
+import MDXContent from '@/components/blogs/mdxContent/MDXContent';
 import { BLOG_CONTENTS_DIR } from '@/constants/blogPosts';
 import getBlogPost from '@/utils/getBlogPost';
 import styles from './page.module.scss';
